Updated Configuration Settings to give ENABLE_ISLAND_SPAWN a CONFIG_NAME, which was causing some errors when attempting to load the config file.
Filled in FloatingIslandGenerator so that Islands do not generate. Updated GeneratorCommon with another helper function and changed canSpawn to take the Random which is passed as an argument to generate, to correct for null pointer exceptions, and hopefully more wisely manage memory.

Updated the mods main file to enable Island Generation by regiestering the FloatingIslandGenerator to the WorldGen Regiestry.

Currently the mod is in a working state, but generation is a little buggy, and occurrs is some odd places... But it does generate Floating Islands, after a fashion...